src/app/github.service.spec.ts
import { TestBed } from '@angular/core/testing';

import { GithubService } from './github.service';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { environment } from '../environments/environment';

describe('GithubService', () => {
  let service: GithubService;
  let httpMock: HttpTestingController;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule]
    });
    httpMock = TestBed.inject(HttpTestingController);
    service = TestBed.inject(GithubService);
  });

  afterEach(() => {
    httpMock.verify();
  });


  it('should be created', () => {
    expect(service).toBeTruthy();
  });

  it('should return User from api', (done: DoneFn) => {
    service.getUser().subscribe((user) => {
      expect(user).not.toBeNull();
      done();
    });

    const request = httpMock.expectOne(environment.apiUrl);
    request.flush(null);
  });
});
